Generalize a base ReCollect Waste entity
